Added Abilities, Respawning, and some other small changes
Base Character class now has the framework for networked abilities setup. The PerformAbility event is implemented in blueprint and can be changed to have a different ability per pawn. The default pawn has Dallas' SNAP.

Dallas SNAP: Reveals all enemies within a certain distance. Done by toggling on their custom render depth which allows the material on them to show the outline. All of this can be seen in the DALLAS folder of blueprints.

Respawning: Simple instant respawn on death, doesn't currently look for a safe place to spawn or anything like that. Might add in the future.